Remove Factorial binary and Fix Makefile
On second thought, even though the example is about using the Factorial
binary with the lldb debugger, the binary should not be committed.

The example showed a .dSYM file being created, but that is not created
with the Swift 2.2 release compiler so make clean fails. Replace *.dSYM
with the Factorial binary.